fix workspace_provider errors on load
